# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WChat 即时通讯系统后端 API 文档",
        "title": "WChat API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页获取指定会话的历史消息，按发送时间倒序返回",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "获取消息列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 UUID",
                        "name": "session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页码 (从1开始)",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / Token 无效",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/messages/{uuid}/recall": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "撤回自己发送的消息（将消息状态设为 2=已撤回），超时可能被拒绝",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "撤回消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "消息 UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "消息不存在 / 超时 / 非本人消息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/sessions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页获取当前用户的聊天会话列表，包含对方名称、头像、最后消息、未读数等",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会话"
                ],
                "summary": "获取会话列表",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页码 (从1开始)",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / Token 无效",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "与目标用户或群组创建一个新会话（如果已存在则返回已有会话的 UUID）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会话"
                ],
                "summary": "创建会话",
                "parameters": [
                    {
                        "description": "target_id=对方UUID, session_type: 0=私聊, 1=群聊",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateSessionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "不能与自己发起会话 / 参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/sessions/{uuid}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "从当前用户的会话列表中删除指定会话",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会话"
                ],
                "summary": "删除会话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "会话不存在",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/sessions/{uuid}/read": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "将指定会话的未读消息数归零",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会话"
                ],
                "summary": "标记会话已读",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "会话不存在",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/sessions/{uuid}/top": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "设置指定会话的置顶状态：1=置顶, 0=取消置顶",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会话"
                ],
                "summary": "置顶/取消置顶会话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "is_top: 0=取消置顶, 1=置顶",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.PinSessionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.CreateSessionReq": {
            "type": "object",
            "required": [
                "target_id"
            ],
            "properties": {
                "session_type": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "target_id": {
                    "type": "string"
                }
            }
        },
...
        "types.PinSessionReq": {
            "type": "object",
            "required": [
                "is_top"
            ],
            "properties": {
                "is_top": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}